tests/requirements.txt
appium-python-client>=3.1.0
pytest>=8.0.0
pytest-html>=4.1.0
pytest-ordering>=0.6
